requirements.txt
epicscorelibs>=7.0.7.99.0.2
epicsdbbuilder>=1.5
numpy
pvxslibs>=1.2.4
setuptools-dso==2.10a1
softioc>=4.4.0
PyYAML~=6.0
pyserial